move SQLiteDatabaseFactory to separate file
