+++ FeedStoreChallenge/SQLiteFeedStore/SQLiteFeedStore.swift
@@ -9,18 +9,6 @@
 import Foundation
 import SQLite
 
-public struct SQLiteDatabaseFactory {
-    
-    private static let dbFolderName = "SQLiteFeedStore"
-    
-    public static func create() -> Connection? {
-        let documentsPath = NSSearchPathForDirectoriesInDomains(.documentDirectory, .userDomainMask, true)[0]
-        let dbPath = "\(documentsPath)/\(dbFolderName)"
-        return try? Connection(dbPath, readonly: false)
-    }
-    
-}
-
 final public class SQLiteFeedStore: FeedStore {
     
     private let db: Connection
